# Conflicts:
#	Framework/Framework.WebApi/Filters/RequiresParametersFilter.cs
#	Social.Application/Dto/Conversation/ConversationSearchDto.cs
#	Social.Application/Dto/Filter/FilterConditionCreateDto.cs
#	Social.Application/Dto/Filter/FilterCreateDto.cs
#	Social.Application/Dto/Filter/FilterUpdateDto.cs
#	Social.WebApi/Controllers/ConversationsController.cs
#	Social.WebApi/Social.WebApi.csproj

+++ Social.Application/AppServices/FilterAppService.cs
@@ -5,6 +5,7 @@ using Social.Application.Dto;
 using Social.Domain;
 using Social.Domain.DomainServices;
 using Social.Domain.Entities;
+using Social.Infrastructure;
 using System;
 using System.Collections.Generic;
 using System.Data.Entity;
@@ -23,6 +24,7 @@ namespace Social.Application.AppServices
         FilterDetailsDto Insert(FilterCreateDto createDto);
         void Delete(int id);
         FilterDetailsDto Update(int id, FilterUpdateDto updateDto);
+        List<FilterManageDto> Sorting(IList<FilterSortDto> dtoList);
     }
 
 
@@ -30,14 +32,17 @@ namespace Social.Application.AppServices
     {
         private IFilterService _domainService;
         private IAgentService _agentService;
+        private INotificationManager _notificationManager;
 
         public FilterAppService(
             IFilterService domainService,
-            IAgentService agentService
+            IAgentService agentService,
+            INotificationManager notificationManager
             )
         {
             _domainService = domainService;
             _agentService = agentService;
+            _notificationManager = notificationManager;
         }
 
         public List<FilterListDto> FindAll()
@@ -58,6 +63,10 @@ namespace Social.Application.AppServices
         public FilterDetailsDto Find(int id)
         {
             var filter = _domainService.Find(id);
+            if (filter == null)
+            {
+                throw SocialExceptions.FilterNotExists(id);
+            }
             var filterDto = Mapper.Map<FilterDetailsDto>(filter);
 
             return filterDto;
@@ -74,29 +83,66 @@ namespace Social.Application.AppServices
             filter = _domainService.Insert(filter);
             CurrentUnitOfWork.SaveChanges();
 
+            _notificationManager.NotifyNewPublicFilter(filter.SiteId, filter.Id);
+
             return Mapper.Map<FilterDetailsDto>(filter);
         }
 
         public void Delete(int id)
         {
+            var filter = _domainService.Find(id);
+            if (filter == null)
+            {
+                throw SocialExceptions.FilterNotExists(id);
+            }
             _domainService.Delete(id);
+            _notificationManager.NotifyNewPublicFilter(filter.SiteId, filter.Id);
         }
 
         public FilterDetailsDto Update(int id, FilterUpdateDto updateDto)
         {
             var updateFilter = _domainService.Find(id);
+            if (updateFilter == null)
+            {
+                throw SocialExceptions.FilterNotExists(id);
+            }
 
             _domainService.DeleteConditons(updateFilter);
             Mapper.Map(updateDto, updateFilter);
             _domainService.UpdateFilter(updateFilter, Mapper.Map<List<FilterConditionCreateDto>, List<FilterCondition>>(updateDto.Conditions.ToList()).ToArray());
             CurrentUnitOfWork.SaveChanges();
 
+            _notificationManager.NotifyNewPublicFilter(updateFilter.SiteId, updateFilter.Id);
             return Mapper.Map<FilterDetailsDto>(updateFilter);
         }
 
         public List<FilterManageDto> FindManageFilters()
         {
             List<Filter> filters = _domainService.FindAll().Where(u => u.IfPublic == true || u.CreatedBy == UserContext.UserId).ToList();
+            List<FilterManageDto> filterDtos = new List<FilterManageDto>();
+            foreach (var filter in filters)
+            {
+                var filterDto = Mapper.Map<FilterManageDto>(filter);
+                filterDto.CreatedByName = _agentService.GetDiaplyName(filter.CreatedBy);
+                filterDtos.Add(filterDto);
+            }
+            return filterDtos;
+        }
+
+        public List<FilterManageDto> Sorting(IList<FilterSortDto> dtoList)
+        {
+            var ids = dtoList.Select(t => t.Id).ToList();
+            List<Filter> filters = _domainService.FindAll().Where(t => ids.Contains(t.Id)).ToList();
+            foreach (var filter in filters)
+            {
+                var dto = dtoList.FirstOrDefault(t => t.Id == filter.Id);
+                if (dto != null)
+                {
+                    filter.Index = dto.Index;
+                    _domainService.Update(filter);
+                }
+            }
+
             List<FilterManageDto> filterDtos = new List<FilterManageDto>();
             foreach (var filter in filters)
             {
